template: Expose complicated sound interface to JS apps
This adds a global "Sounds" object which can be used from inside the JS
apps.

For async sounds, we need to get the return value (the UUID) back into JS
so we can later stop the sound if necessary. This is difficult, since
WebKit's postMessage does not allow you to get a return value. The
logical way would be to add a callback to the message arguments, but
that will also not work, as functions are not allowed to be sent across
that security boundary. (You get a DataCloneError if you try.) This code
takes a function, stringifies it to get its source code, and then when
the UUID is returned it will eval the function in the window's global
context.

This approach is not going to be used because it is awful, but I can
anticipate that we might need it anyway in the future. Since it was hard
to get right, I am leaving it intact in this commit and building a
better interface in the next commit.

https://phabricator.endlessm.com/T24430