template: play and stop loops robustly
Loops don't stop playing by their own, like other sounds,
therefore, we need to make sure we track those, e.g.  if
the game logic is broken and adds the same loop non-stop
there is literally no way to stop the previous instances
of that loop sound.

So, if we play a loop sound, we must stop any previous
instance of that same sound, so we don't lose track.

https://phabricator.endlessm.com/T24791