HackUnlock: Adapt UI to different resolutions
To be able to resolve this puzzle the background should be the same size
for all resolutions, in other case, the wave won't match the background.

The background is centered, with a fixed size of 1920x1040, independent
of the window size and we do the same for the wave.

If the window is bigger than the fixed size of 1920x1040, we fill the
offset with a black background.

https://phabricator.endlessm.com/T24712