LightSpeed: Keep track of min and max Y position for every enemy
As per Noel, the clubhouse watches these global parameters for min and
max Y position for every type of enemy, in order to determine whether
certain quests are solved. We generalize the existing parameter for
obstacle type 1. The parameters are reset at the beginning of every
level (I changed the initial values to ±Infinity to avoid any problems
in case of different resolutions or a different coordinate system.)

https://phabricator.endlessm.com/T25508