Lightspeed: Save all level parameters, not just the current level
This changes the format of the saved state so that all parameters of all
levels are saved. Previously, only the parameters of the current level
were saved, so once you finished a level you'd wipe out everything you
had changed in any previous levels.

Since there is now no level dependency in the saved state, we move the
saveState method to be a global function instead.

https://phabricator.endlessm.com/T25561