Fizzics: Clean up levels JSON
JSON supports integers, so we can remove the hundreds of expensive
parseInt() calls every time we load a level.

There is existing code to load collision test parameters from the JSON
file, but these parameters are never given in the existing JSON file. We
will do something with this in a subsequent commit, but for now make
sure the parameters end up as NaN for backwards compatibility (which
would be the output of `parseInt(undefined)` which is what ends up
happening in the existing code.)

https://phabricator.endlessm.com/T26026