Fizzics: Make level parameters easier to read
Most of the level parameters are duplicated. Here we create a default
level parameters object that is basically a blank level. We also create
a second default for levels that are part of the "main game", with the
pool table background and the expected behaviour of the cricket balls,
gems, and spiky balls. The quest levels are based on the blank defaults.

This is in preparation for moving all non-default level parameters into
the levels JSON file.

We have to specify the parameters for species 4 (gems) in the QUEST1
level, since the previous behaviour relied on a main level having been
set first, and just didn't change the species 4 parameters.

Some of the special levels specified a deathEffect parameter, but this
parameter doesn't seem to be used anywhere in the game, so we ignore it.

https://phabricator.endlessm.com/T26026