Fizzics: Move quest level presets into levels JSON
When levels have parameters that are different from the defaults, we now
specify those in the levels JSON file, in a "preset" key. That makes
sense because then all the information needed to recreate a level is in
the JSON file.

The main levels get "preset": {}, and the quest levels, which are now
included in the JSON file, get the data that was previously specified in
code, in setLevelGlobalParams().

The special quest levels currently have all their balls specified in
code. Since the existing code depends on the canvas width and height, we
don't move this to the JSON file. Instead, we give the quest levels
"balls": [] in the JSON file, and move the quest levels populating code
into a new function _placeBallsProgrammatically().

We have to change a few things because there was code relying on the
length of the levels array in the JSON file. Mostly the purpose of such
code was to determine whether a level was a main level or a quest level.
In these cases we check if the level is < QUEST0, as is already done in
other places in the code.

There was also existing code that loaded the collision test parameters
from the JSON file, but those parameters were never specified in the
JSON file, instead they were given in code, in the same place where the
balls were placed programmatically. We move these to the JSON file as
well.

https://phabricator.endlessm.com/T26026