Fizzics: Round ball positions when saving
This removes the "generation loss" from the level editor. Balls are
still placed onto the field with a certain amount of jitter, but we
round to the nearest integer pixel when saving so that the jitter isn't
saved along with the level and compounded every time the level is loaded
and saved.

https://phabricator.endlessm.com/T26161