Only assign levelParameters once
For some reason this was causing a long (but not infinite) notification
loop in Sidetrack, which often ended with one or both of the code
variables being empty. I did not figure out why the loop occurred, but
assigning the properties of levelParameters[i] only once does solve it.

Same change made in Lightspeed. This is good practice anyway, because
each assignment calls a setter, so we don't really want the setter to be
called with a value that's about to be overwritten anyway.

https://phabricator.endlessm.com/T26381