Sidetrack: Add cutscene quest variables
These variables will play the cutscenes when set to true, and then set
themselves back to false to notify the quest script when the cutscene is
finished.

For now, since we don't have cutscenes, we just wait 5 seconds and then
set them to false.

For the controls cutscene, we manually advance to level 14 and unlock
the rest of the levels. In reality, the quest should do this.

https://phabricator.endlessm.com/T26378