Avoid race condition when loading state
It's possible that the main game scene starts (with default values for
the game state parameters) before the game state finishes loading, in
which case the game state is basically ignored.

This adds a gate to the initial scenes of both Sidetrack and Lightspeed
so that they do not proceed into the main game scene until both of these
conditions are satisfied:

- Assets are preloaded (Sidetrack) / Player is ready (Lightspeed)
- The game state has been loaded.

This bug is much less likely in Lightspeed since the player will nearly
always be slower to click the start button than the game state will take
to load, but it's good practice anyway.

https://phabricator.endlessm.com/T26585