Separate loading game state out from ToyApp.loadNotify()
Previously, apps that opted in to splashscreen notification would
receive their game state in a callback which would be called _after_
notifying the runner that they were ready for the splashscreen to be
taken down with ToyApp.loadNotify().

This is no longer sufficient, as we want to go directly into the level
where the player is, and to know which level that is, we need to load
the game state. So without this change, we will always get a flash of
nothing after the splashscreen is taken down but before the level
starts, while the game state is being fetched.

We separate the two actions out; now apps must request their game state
to be loaded separately from ToyApp.loadNotify(). We add a new
ToyApp.requestState() method for this, and call it as early as possible.

Sidetrack is currently the only app that this applies to, because
Lightspeed is the only other app that uses the game state service, and
Lightspeed does not use the splashscreen notification.

Additionally, in Sidetrack, we delay the call of ToyApp.loadNotify()
until the level is actually started, so that we avoid the flash. In
Lightspeed this isn't important, as we can show the title scene before
the game state has been loaded.

https://phabricator.endlessm.com/T26585