Fizzics: Don't allow creating species if it is disabled
Ideally we would listen to changes on this parameter and select the
first species that it is allowed to create, but a notification mechanism
doesn't exist in Fizzics.

Therefore we check the global parameters when adding a ball, and play
the "BZZT!" sound if that ball's species is not allowed to be added.

https://phabricator.endlessm.com/T26391