Fizzics: Fix positioning of the top board on resize
The top board gets assigned a fixed position that is the size of the
window at the time, so when resizing the window or when it is resized
while the game is getting initialized (happens sometimes), then the top
board's position is updated but uses the initial values which could be
already scaled.

This patch fixes that by using the expected original width for the
top board's position. So the board will be scaled correctly when the
window is resized (otherwise, if the scaled width is used, the position
will not be the middle of the screen).

https://phabricator.endlessm.com/T26164