Fizzics: Fix positioning the balls when the window is resized
When the window is resized, the balls get resized correctly. However,
if the game is reset, then the balls will not be positioned in the
correct places because the right/bottom walls are used with scaled
values, so the balls' positions seem to end up getting scaled twice.

The same issue happens to a lot of UI elements if the window is
resized while the game is getting initialized: this can be reproduced
by launching Fizzics and quickly pressing the Super key to go to the
overview (when Fizzics is finally launched, many UI elements are
incorrectly positioned).

By using the original width/height values for the right/bottom walls
the UI/balls are be placed correctly even when the size is different
as they get correctly scaled.

https://phabricator.endlessm.com/T26164