Fizzics: Fix top board positioning
In a previous commit (9f9112e) it was supposed to pass the default width
to the UI initialization function, but due to a mistake, the part where
the function used that value was not added to the commit.

This patch adds that missing part, thus fixing the positioning of the
top board when the window is resized while being initialized (click on
the Fizzics icon and when the splash screen is shown hit the Super key
to show the overview).

https://phabricator.endlessm.com/T26164