Fizzics: Only check if adding balls is disabled when it's user started
The createBall function was checking whether the ball species creation
had been disabled, and this meant that when resetting the level after
one or more balls had been disabled (as in the Episode4 FizzicsKey
quest) the ball repositioning didn't work.

This patch fixes that by moving the "disabled-ball-creation" check into
the functions that are triggered by the user's actions.

https://phabricator.endlessm.com/T26698