Fizzics: Add canvas container (changes resizing)
To embed the toy app in another web application, we need to specify a
container. By default, Phaser adds a canvas to the document body. This
change makes Phaser look for a parent with ID `toy-app-container`.

Warning: This changes the resizing logic. In desktop, Fizzics scales
to fit the initial size of the window. Then it keeps the aspect ratio
when the window is resized.