fizzics: Map collision sprites to index set in the toolbox
In HackToolbox, deathVisualGood and deathVisualBad were mapping
to a specific index. In the phaser port, collision images have been
stored with a different index. So for example, selecting a confetti
visual effect in the toolbox, would reflect in the the phaser app
in a wrong collision effect, for example, a explosion effect.

This commit also changes defaults deathVisualGood constants so
the collision effect "main - diamond" and "main - star" are
consistent with Fizzics 1.

https://phabricator.endlessm.com/T29530