fizzics: Display main ball effect on star ball collision
When changing the variable deathVisualGood_1 (the effect that
activates when a main ball goodly dies) from whether from the
web or the toolbox, no changes were apparentely applied when the
main ball.

For example, if in the toolbox of the desktop application, in the
tab of the main orange ball, you change in the select box of the
Good visual effect to "Skull", and then you flip back the app,
when the main ball collided with the main ball the confetti (the
deafult effect) was still displayed, instead of the new skull
effect.

https://phabricator.endlessm.com/T29530