fizzics: Support attraction force between balls
The attractors plugin have been enabled to apply attraction force between
balls. A radius where this attraction force works is applied.

gravity is set to be ignored when setting the gravityScale to 0,
because otherwise the attraction force does not work. It seems that
gravity scale is multiplied by the resultant attraction force, so if this
scale is 0, the attraction force is 0. That is something to confirm that.
To workaround this issue for now, we set to ignore the gravity in this case.

TODO: Prevent attraction force when flinging the ball. In this case, we
should check in the `doAttraction` function that FlingerView.state is

https://phabricator.endlessm.com/T29346