Sidetrack: Make it functional without app bundle
This patch adds some links to the needed common resources inside the app
to make it functional without the flatpak bundle so we can use this toy
app in the hack-web.

The globalParameters check is removed in main.js because the hack-web
creates a Proxy and the variable won't be the same, in any case this
change shouldn't affect the app behavior because the startLevel property
is only present in the globalParameters.

https://phabricator.endlessm.com/T29727