Sidetrack: Don't check globalParameters with Object.is
In the gameScene we are testing if a callback is done because the
globalParameters changes and to do that the object is compared with the
original one, but in the hack-web we're creating another Proxy on top of
that so this check always fails.

This patch just check if the object has the 'currentLevel' property so
this can continue working in the desktop version and in hack web.

https://phabricator.endlessm.com/T29826