Rename sidetrack and OS apps to include in clubhouse
These apps are not usable without the clubhouse so instead of release as
independent apps we can embed inside the clubhouse bundle flatpak and we
only need to release one flatpak with the clubhouse and toy apps.

This patch renames the apps to make them independent app inside the same
clubhouse flatpak. We need to use the same base name to be able to
export more than one .desktop files.

https://phabricator.endlessm.com/T30458